updated resources
