--- src/main/java/co/crisi/helidon/odyssey/jotvault/service/UserService.java
@@ -1,6 +1,0 @@
-package co.crisi.helidon.odyssey.jotvault.service;
-
-import co.crisi.helidon.odyssey.jotvault.model.IUser;
-
-public interface UserService extends CrudProvider<IUser, Long> {
-}
